refactor(selectors): consolidate both color selectors into one; move selectors into root components dir; move selector styles into same dir as components;
