+++ src/components/selectors/index.ts
@@ -1,0 +1,3 @@
+export * from "./color-selector"
+export * from "./size-selector"
+export * from "./selector"
