+++ src/styles/index.ts
@@ -1,4 +1,3 @@
 export * from "./primitives";
 export { default as card } from "./product-card";
-export * from "./selector";
 export * from "./typography";
